+++ src/WebFlux/Extensions/ServiceCollectionExtensions.cs
@@ -3,7 +3,6 @@ using Microsoft.Extensions.Configuration;
 using Microsoft.Extensions.DependencyInjection;
 using Microsoft.Extensions.DependencyInjection.Extensions;
 using Microsoft.Extensions.Logging;
-using WebFlux.Configuration;
 using WebFlux.Core.Interfaces;
 using WebFlux.Core.Models;
 using WebFlux.Services;